requirements.txt
greenlet==0.4.14
msgpack==0.5.6
neovim==0.2.6
pycairo==1.17.1
PyGObject==3.28.3
